fix: use streaming retry settings for ResumableStreamIterator
